tests/wpt/web-platform-tests/websockets/Create-on-worker-shutdown.html

<!DOCTYPE html>
<html>
<head>
    <title>W3C WebSocket API - Create WebSocket - on a worker after self.close()</title>
    <script type="text/javascript" src="/resources/testharness.js"></script>
    <script type="text/javascript" src="/resources/testharnessreport.js"></script>
    <script type="text/javascript" src="websocket.sub.js"></script>
</head>
<body>
    <div id="log"></div>
    <script type="text/javascript">

async_test(t => {
  function workerCode() {
    close();
    var ws = new WebSocket(self.location.origin.replace('http', 'ws'));
    postMessage(ws.readyState == WebSocket.CONNECTING);
  }

  var workerBlob = new Blob([workerCode.toString() + ";workerCode();"], {type:"application/javascript"});

  var w = new Worker(URL.createObjectURL(workerBlob));
  w.onmessage = function(e) {
    assert_true(e.data, "WebSocket created on worker shutdown.");
    t.done();
  }
}, 'WebSocket created after a worker self.close()');

    </script>
</body>
</html>
